+++ built/grammar/GrammarCustomConstructorsNode.d.ts
@@ -1,0 +1,14 @@
+import TreeNode from "../base/TreeNode";
+import types from "../types";
+declare abstract class AbstractCustomConstructorNode extends TreeNode {
+    private _getBuiltInConstructors;
+    getDefinedConstructor(): types.RunTimeNodeConstructor;
+    protected _getCustomConstructor(): types.RunTimeNodeConstructor;
+    getErrors(): types.ParseError[];
+    getBuiltIn(): any;
+}
+declare class GrammarCustomConstructorsNode extends TreeNode {
+    getKeywordMap(): {};
+    getConstructorForEnvironment(): AbstractCustomConstructorNode;
+}
+export default GrammarCustomConstructorsNode;